refactor: add subcomponents folders to view to tidy patterns folder
