+++ src/components/views/Summary/Summary.tsx
@@ -1,11 +1,11 @@
 import { Box, Container, Typography } from "@mui/material";
-import bgImage from "../../images/backgrounds/wave.svg";
-import { backgroundImg, flexCenter } from "../../utility/styling";
-import List from "../patterns/List/List";
-import Progress from "../patterns/Progress/Progress";
-import HighlitedText from "../atoms/HighlitedText/HighlitedText";
-import { useAnswersContext } from "../../context/Answers/Answers";
-import type { AnsweredQuestion } from "../../types/globalTypes";
+import bgImage from "../../../images/backgrounds/wave.svg";
+import { backgroundImg, flexCenter } from "../../../utility/styling";
+import List from "./subcomponents/List";
+import Progress from "../../patterns/Progress/Progress";
+import HighlitedText from "../../atoms/HighlitedText/HighlitedText";
+import { useAnswersContext } from "../../../context/Answers/Answers";
+import type { AnsweredQuestion } from "../../../types/globalTypes";
 
 function calculateOutcome(questions: AnsweredQuestion[]) {
   let points = 0;
